refactor: change files for kebab-case
