+++ src/useCases/register-message-use-case.spec.ts
@@ -1,4 +1,4 @@
-import { RegisterMessageUseCase } from "./registerMessageUseCase"
+import { RegisterMessageUseCase } from "./register-message-use-case"
 
 const createMessageSpy = jest.fn();
